+++ web/ui/react-app/src/App.css
@@ -243,10 +243,6 @@ div.time-input {
     display: table;
 }
 
-.label-component {
-    margin-bottom: 1%;
-}
-
 .btn-sd-style {
     padding: 1px;
     font-size: 10px;
